test: add initial basic tests
