+++ .github/workflows/format.yml
@@ -12,6 +12,7 @@ concurrency:
 jobs:
   format:
     if: github.repository == 'remix-run/release-comment-action'
+    name: 👔 Format
     runs-on: ubuntu-latest
 
     steps:
