+++ .github/workflows/no-response.yml
@@ -13,6 +13,7 @@ jobs:
   stale:
     if: github.repository == 'remix-run/release-comment-action'
     runs-on: ubuntu-latest
+    name: 🥺 Handle Ghosting
     steps:
       - name: 🥺 Handle Ghosting
         uses: actions/stale@v8
