+++ .github/workflows/test.yml
@@ -1,4 +1,4 @@
-name: 🔬 Lint
+name: 🧪 Test
 
 on:
   push:
@@ -11,9 +11,8 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  lint:
-    if: github.repository == 'remix-run/release-comment-action'
-    name: 🔬 Lint
+  test:
+    name: 🧪 Test
     runs-on: ubuntu-latest
     steps:
       - name: ⬇️ Checkout repo
@@ -27,5 +26,5 @@ jobs:
             - recursive: true
               args: [--frozen-lockfile, --strict-peer-dependencies]
 
-      - name: 🔬 Lint
-        run: pnpm run lint
+      - name: 🧪 Test
+        run: pnpm run test
